agent-as-tool/pyproject.toml
[project]
name = "tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.2",
    "python-dotenv>=1.1.1",
]
